Splitt ut komponenter i egne filer
